Stability model: integrate with ml_neural.py::NeuralNetwork class, combine data generation with training
The purpose of this change is to create a model where just a *.pt file is needed to predict whether the input F4 is stable or unstable. The training should be performed by creating a folder inside stability_model and copying the train_model.py file. plot.py will plot the test-train loss once the training finishes.